pyproject.toml
[project]
name = "rag-observability"
version = "0.1.0"
description = "Local RAG app with observability"
requires-python = ">=3.10"

dependencies = [
    "fastapi",
    "uvicorn",
    "chromadb",
    "python-multipart",
    "pymupdf",
    "httpx",
    "loguru",
    "sqlite-utils",
    "streamlit",
    "pydantic-settings",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-exporter-otlp",
    "opentelemetry-instrumentation-fastapi",
    "opentelemetry-instrumentation-httpx",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-mock",
]

[tool.setuptools]
packages = ["app", "dashboard"]

[tool.uv]
package = true
